+++ src/Migrator.Framework/Loggers/Logger.cs
@@ -45,32 +45,32 @@ namespace Migrator.Framework.Loggers
 			_writers.Remove(writer);
 		}
 
-		public void Started(int currentVersion, int finalVersion)
+		public void Started(long currentVersion, long finalVersion)
 		{
 			WriteLine("Current version : {0}", currentVersion);
 		}
 
-		public void MigrateUp(int version, string migrationName)
+		public void MigrateUp(long version, string migrationName)
 		{
 			WriteLine("{0} {1}", version.ToString().PadLeft(_widthFirstColumn), migrationName);
 		}
 
-		public void MigrateDown(int version, string migrationName)
+		public void MigrateDown(long version, string migrationName)
 		{
 			MigrateUp(version, migrationName);
 		}
 
-		public void Skipping(int version)
+		public void Skipping(long version)
 		{
 			WriteLine("{0} {1}", version.ToString().PadLeft(_widthFirstColumn), "<Migration not found>");
 		}
 
-		public void RollingBack(int originalVersion)
+		public void RollingBack(long originalVersion)
 		{
 			WriteLine("Rolling back to migration {0}", originalVersion);
 		}
 
-		public void Exception(int version, string migrationName, Exception ex)
+		public void Exception(long version, string migrationName, Exception ex)
 		{
 			WriteLine("{0} Error in migration {1} : {2}", "".PadLeft(_widthFirstColumn), version, ex.Message);
 			if (_trace)
@@ -89,7 +89,7 @@ namespace Migrator.Framework.Loggers
 			}
 		}
 
-		public void Finished(int originalVersion, int currentVersion)
+		public void Finished(long originalVersion, long currentVersion)
 		{
 			WriteLine("Migrated to version {0}", currentVersion);
 		}
